Move MilkdropNoise class to Renderer directory.
It's only referenced from within TextureManager, plus we can reuse the noise generator for other things.
